Fixing model to match apidata
